\name{categorize}
\alias{categorize}
\title{
  Categorization of a continuous variable.
}
\description{
  Internaly used function to categorize a continuous variable.
}
\usage{
categorize(dat, threshold)
}
\arguments{
  \item{dat}{
    Numerical vector for a categorization.
  }
  \item{threshold}{
    Vector or scalar value(s) to determine the cutoff values for a categorization.
  }
}
